+++ packages/core/styles/src/properties/base.ts
@@ -44,10 +44,12 @@ export const baseColors = {
   "neutral-textLow": varsThemeBase.colors.neutral.textLow,
   "neutral-textHigh": varsThemeBase.colors.neutral.textHigh,
 
-  "ai-generativeSurface": varsThemeBase.colors.aiGenerative.surface,
-  "ai-generativeSurfaceHighlight":
+  "aiGenerative-background": varsThemeBase.colors.aiGenerative.background,
+  "aiGenerative-surface": varsThemeBase.colors.aiGenerative.surface,
+  "aiGenerative-surfaceHighlight":
     varsThemeBase.colors.aiGenerative.surfaceHighlight,
-  "ai-generativeInteractive": varsThemeBase.colors.aiGenerative.background,
-  "ai-generativeTextLow": varsThemeBase.colors.aiGenerative.textLow,
-  "ai-generativeTextHigh": varsThemeBase.colors.aiGenerative.textHigh,
+  "aiGenerative-textLow": varsThemeBase.colors.aiGenerative.textLow,
+  "aiGenerative-textHigh": varsThemeBase.colors.aiGenerative.textHigh,
+  "aiGenerative-interactivePressed":
+    varsThemeBase.colors.aiGenerative.interactivePressed,
 };